chore(test): restructure helpers and deduplicate some functions
