+++ test/internal/helpers/aws/testCases.go
@@ -1,0 +1,1 @@
+package awsHelpers
