+++ test/multi_region_aws_operational_procedure_test.go
@@ -1,17 +1,15 @@
 package multiregionawsoperationalprocedure
 
 import (
-	"context"
 	"fmt"
-	"os"
-	"os/exec"
 	"strings"
 	"testing"
 	"time"
 
 	"multiregiontests/internal/helpers"
+	awsHelpers "multiregiontests/internal/helpers/aws"
+	kubectlHelpers "multiregiontests/internal/helpers/kubectl"
 
-	"github.com/camunda/zeebe/clients/go/v8/pkg/zbc"
 	"github.com/gruntwork-io/terratest/modules/k8s"
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/require"
@@ -40,59 +38,12 @@ var secondary helpers.Cluster
 
 func TestSetupTerraform(t *testing.T) {
 	t.Log("[TF SETUP] Applying Terraform config 👋")
-
-	terraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
-		TerraformDir: terraformDir,
-		Vars: map[string]interface{}{
-			"cluster_name": clusterName,
-			"aws_profile":  awsProfile,
-		},
-		NoColor: true,
-	})
-
-	terraform.InitAndApply(t, terraformOptions)
-
-	t.Log("[TF SETUP] Generating kubeconfig files 📜")
-
-	cmd := exec.Command("aws", "eks", "--region", "eu-west-3", "update-kubeconfig", "--name", fmt.Sprintf("%s-paris", clusterName), "--profile", awsProfile, "--kubeconfig", "kubeconfig-paris")
-
-	_, err := cmd.Output()
-	if err != nil {
-		t.Fatalf("[TF SETUP] could not run command: %v", err)
-		return
-	}
-
-	require.FileExists(t, "kubeconfig-paris", "kubeconfig-paris file does not exist")
-
-	cmd2 := exec.Command("aws", "eks", "--region", "eu-west-2", "update-kubeconfig", "--name", fmt.Sprintf("%s-london", clusterName), "--profile", awsProfile, "--kubeconfig", "kubeconfig-london")
-
-	_, err2 := cmd2.Output()
-	if err2 != nil {
-		t.Fatalf("[TF SETUP] could not run command: %v", err2)
-		return
-	}
-
-	require.FileExists(t, "kubeconfig-london", "kubeconfig-london file does not exist")
+	awsHelpers.TestSetupTerraform(t, terraformDir, clusterName, awsProfile)
 }
 
 func TestTeardownTerraform(t *testing.T) {
 	t.Log("[TF TEARDOWN] Destroying workspace 🖖")
-
-	terraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
-		TerraformDir: terraformDir,
-		Vars: map[string]interface{}{
-			"cluster_name": clusterName,
-			"aws_profile":  awsProfile,
-		},
-		NoColor: true,
-	})
-	terraform.Destroy(t, terraformOptions)
-
-	os.Remove("kubeconfig-paris")
-	os.Remove("kubeconfig-london")
-
-	require.NoFileExists(t, "kubeconfig-paris", "kubeconfig-paris file still exists")
-	require.NoFileExists(t, "kubeconfig-london", "kubeconfig-london file still exists")
+	awsHelpers.TestTeardownTerraform(t, terraformDir, clusterName, awsProfile)
 }
 
 // AWS EKS Multi-Region Tests
@@ -101,9 +52,9 @@ func TestAWSOperationalProcedure(t *testing.T) {
 	t.Log("[2 REGION TEST] Running tests for AWS EKS Multi-Region 🚀")
 
 	// For CI run it separately
-	// go test --count=1 -v -timeout 120m ../test -run TestSetupTerraform
-	// go test --count=1 -v -timeout 120m ../test -run Test2RegionAWSEKS
-	// go test --count=1 -v -timeout 120m ../test -run TestTeardownTerraform
+	// go test --count=1 -v -timeout 120m ./multi_region_aws_operational_procedure_test.go -run TestSetupTerraform
+	// go test --count=1 -v -timeout 120m ./multi_region_aws_operational_procedure_test.go -run TestAWSOperationalProcedure
+	// go test --count=1 -v -timeout 120m ./multi_region_aws_operational_procedure_test.go -run TestTeardownTerraform
 
 	// Pre and Post steps - deactivated for CI
 	// setupTerraform(t)
@@ -114,6 +65,7 @@ func TestAWSOperationalProcedure(t *testing.T) {
 		name  string
 		tfunc func(*testing.T)
 	}{
+		// AWS Multi Region Setup
 		{"TestInitKubernetesHelpers", initKubernetesHelpers},
 		{"TestClusterReadyCheck", clusterReadyCheck},
 		{"TestCrossClusterCommunication", testCrossClusterCommunication},
@@ -124,6 +76,7 @@ func TestAWSOperationalProcedure(t *testing.T) {
 		{"TestDeployC8Helm", deployC8Helm},
 		{"TestCheckC8RunningProperly", checkC8RunningProperly},
 		{"TestDeployC8processAndCheck", deployC8processAndCheck},
+		// Multi-Region Operational Procedure
 		{"TestCheckTheMath", checkTheMath},
 		{"TestDeleteSecondaryRegion", deleteSecondaryRegion},
 		{"TestCreateFailoverDeploymentPrimary", createFailoverDeploymentPrimary},
@@ -158,62 +111,34 @@ func TestAWSOperationalProcedure(t *testing.T) {
 
 func initKubernetesHelpers(t *testing.T) {
 	t.Log("[K8S INIT] Initializing Kubernetes helpers 🚀")
-	primary = helpers.Cluster{
-		Region:           "eu-west-2",
-		ClusterName:      fmt.Sprintf("%s-london", clusterName),
-		KubectlNamespace: *k8s.NewKubectlOptions("", kubeConfigPrimary, "camunda-primary"),
-		KubectlSystem:    *k8s.NewKubectlOptions("", kubeConfigPrimary, "kube-system"),
-		KubectlFailover:  *k8s.NewKubectlOptions("", kubeConfigPrimary, "camunda-primary-failover"),
-	}
-	secondary = helpers.Cluster{
-		Region:           "eu-west-3",
-		ClusterName:      fmt.Sprintf("%s-paris", clusterName),
-		KubectlNamespace: *k8s.NewKubectlOptions("", kubeConfigSecondary, "camunda-secondary"),
-		KubectlSystem:    *k8s.NewKubectlOptions("", kubeConfigSecondary, "kube-system"),
-		KubectlFailover:  *k8s.NewKubectlOptions("", kubeConfigSecondary, "camunda-secondary-failover"),
-	}
-
-	k8s.CreateNamespace(t, &primary.KubectlNamespace, "camunda-primary")
-	k8s.CreateNamespace(t, &primary.KubectlFailover, "camunda-primary-failover")
-	k8s.CreateNamespace(t, &secondary.KubectlNamespace, "camunda-secondary")
-	k8s.CreateNamespace(t, &secondary.KubectlFailover, "camunda-secondary-failover")
+	awsHelpers.InitKubernetesHelpers(t, primary, secondary, kubeConfigPrimary, kubeConfigSecondary, clusterName)
 }
 
 func clusterReadyCheck(t *testing.T) {
 	t.Log("[CLUSTER CHECK] Checking if clusters are ready 🚦")
-	clusterStatusPrimary := helpers.WaitForCluster(primary.Region, primary.ClusterName)
-	clusterStatusSecondary := helpers.WaitForCluster(secondary.Region, secondary.ClusterName)
-
-	require.Equal(t, "ACTIVE", clusterStatusPrimary)
-	require.Equal(t, "ACTIVE", clusterStatusSecondary)
-
-	nodeGroupStatusPrimary := helpers.WaitForNodeGroup(primary.Region, primary.ClusterName, "services")
-	nodeGroupStatusSecondary := helpers.WaitForNodeGroup(secondary.Region, secondary.ClusterName, "services")
-
-	require.Equal(t, "ACTIVE", nodeGroupStatusPrimary)
-	require.Equal(t, "ACTIVE", nodeGroupStatusSecondary)
+	awsHelpers.ClusterReadyCheck(t, primary, secondary)
 }
 
 func testCrossClusterCommunication(t *testing.T) {
 	t.Log("[CROSS CLUSTER] Testing cross-cluster communication with IPs 📡")
-	helpers.CrossClusterCommunication(t, false, k8sManifests, primary, secondary)
+	kubectlHelpers.CrossClusterCommunication(t, false, k8sManifests, primary, secondary)
 }
 
 func applyDnsChaining(t *testing.T) {
 	t.Log("[DNS CHAINING] Applying DNS chaining 📡")
-	helpers.DNSChaining(t, primary, secondary, k8sManifests)
-	helpers.DNSChaining(t, secondary, primary, k8sManifests)
+	awsHelpers.DNSChaining(t, primary, secondary, k8sManifests)
+	awsHelpers.DNSChaining(t, secondary, primary, k8sManifests)
 }
 
 func testCoreDNSReload(t *testing.T) {
 	t.Logf("[COREDNS RELOAD] Checking for CoreDNS reload 🔄")
-	helpers.CheckCoreDNSReload(t, &primary.KubectlSystem)
-	helpers.CheckCoreDNSReload(t, &secondary.KubectlSystem)
+	kubectlHelpers.CheckCoreDNSReload(t, &primary.KubectlSystem)
+	kubectlHelpers.CheckCoreDNSReload(t, &secondary.KubectlSystem)
 }
 
 func testCrossClusterCommunicationWithDNS(t *testing.T) {
 	t.Log("[CROSS CLUSTER] Testing cross-cluster communication with DNS 📡")
-	helpers.CrossClusterCommunication(t, true, k8sManifests, primary, secondary)
+	kubectlHelpers.CrossClusterCommunication(t, true, k8sManifests, primary, secondary)
 }
 
 func deployC8Helm(t *testing.T) {
@@ -221,9 +146,9 @@ func deployC8Helm(t *testing.T) {
 
 	// We have to install both at the same time as otherwise zeebe will not become ready
 
-	helpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, false, false, false, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, false, false, false, map[string]string{})
 
-	helpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, false, false, false, false, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, false, false, false, false, map[string]string{})
 
 	// Check that all deployments and Statefulsets are available
 	// Terratest has no direct function for Statefulsets, therefore defaulting to pods directly
@@ -259,102 +184,18 @@ func deployC8Helm(t *testing.T) {
 
 func checkC8RunningProperly(t *testing.T) {
 	t.Log("[C8 CHECK] Checking if Camunda Platform is running properly 🚦")
-	service := k8s.GetService(t, &primary.KubectlNamespace, "camunda-zeebe-gateway")
-	require.Equal(t, service.Name, "camunda-zeebe-gateway")
-
-	tunnel := k8s.NewTunnel(&primary.KubectlNamespace, k8s.ResourceTypeService, "camunda-zeebe-gateway", 0, 26500)
-	defer tunnel.Close()
-	tunnel.ForwardPort(t)
-
-	client, err := zbc.NewClient(&zbc.ClientConfig{
-		GatewayAddress:         tunnel.Endpoint(),
-		UsePlaintextConnection: true,
-	})
-	if err != nil {
-		t.Fatalf("[C8 CHECK] Failed to create client: %v", err)
-		return
-	}
-
-	defer client.Close()
-
-	// Get the topology of the Zeebe cluster
-	topology, err := client.NewTopologyCommand().Send(context.Background())
-	if err != nil {
-		t.Fatalf("[C8 CHECK] Failed to get topology: %v", err)
-		return
-	}
-
-	require.Equal(t, 8, len(topology.Brokers))
-
-	primaryCount := 0
-	secondaryCount := 0
-
-	t.Log("[C8 CHECK] Cluster status:")
-	for _, broker := range topology.Brokers {
-		if strings.Contains(broker.Host, "camunda-primary") {
-			primaryCount++
-		} else if strings.Contains(broker.Host, "camunda-secondary") {
-			secondaryCount++
-		}
-		t.Logf("[C8 CHECK] Broker ID: %d, Address: %s, Partitions: %v\n", broker.NodeId, broker.Host, broker.Partitions)
-	}
-
-	require.Equal(t, 4, primaryCount)
-	require.Equal(t, 4, secondaryCount)
+	kubectlHelpers.CheckC8RunningProperly(t, primary)
 }
 
 func deployC8processAndCheck(t *testing.T) {
 	t.Log("[C8 PROCESS] Deploying a process and checking if it's running 🚀")
-	service := k8s.GetService(t, &primary.KubectlNamespace, "camunda-zeebe-gateway")
-	require.Equal(t, service.Name, "camunda-zeebe-gateway")
-
-	tunnel := k8s.NewTunnel(&primary.KubectlNamespace, k8s.ResourceTypeService, "camunda-zeebe-gateway", 0, 26500)
-	defer tunnel.Close()
-	tunnel.ForwardPort(t)
-
-	client, err := zbc.NewClient(&zbc.ClientConfig{
-		GatewayAddress:         tunnel.Endpoint(),
-		UsePlaintextConnection: true,
-	})
-	if err != nil {
-		t.Fatalf("[C8 PROCESS] Failed to create client: %v", err)
-		return
-	}
-
-	defer client.Close()
-
-	ctx := context.Background()
-	response, err := client.NewDeployResourceCommand().AddResourceFile(fmt.Sprintf("%s/single-task.bpmn", resourceDir)).Send(ctx)
-	if err != nil {
-		t.Fatalf("[C8 PROCESS] %s", err)
-		return
-	}
-	t.Logf("[C8 PROCESS] Created process: %s", response.String())
-	require.NotEmpty(t, response.String())
-	require.Contains(t, response.String(), "bigVarProcess")
-
-	t.Log("[C8 PROCESS] Sleeping shortly to let process be propagated")
-	time.Sleep(30 * time.Second)
-
-	t.Log("[C8 PROCESS] Starting another Process instance 🚀")
-	msg, err := client.NewCreateInstanceCommand().BPMNProcessId("bigVarProcess").LatestVersion().Send(ctx)
-	if err != nil {
-		t.Fatalf("[C8 PROCESS] %s", err)
-		return
-	}
-	t.Logf("[C8 PROCESS] Created process: %s", msg.String())
-	require.NotEmpty(t, msg.String())
-	require.Contains(t, msg.String(), "bigVarProcess")
-
-	// check that was exported to ElasticSearch and available via Operate
-	helpers.CheckOperateForProcesses(t, primary)
-	helpers.CheckOperateForProcesses(t, secondary)
+	kubectlHelpers.DeployC8processAndCheck(t, primary, secondary, resourceDir)
 }
 
 func teardownAllC8Helm(t *testing.T) {
 	t.Log("[C8 HELM TEARDOWN] Tearing down Camunda Platform Helm Chart 🚀")
-	helpers.TeardownC8Helm(t, &primary.KubectlNamespace)
-	helpers.TeardownC8Helm(t, &secondary.KubectlNamespace)
+	kubectlHelpers.TeardownC8Helm(t, &primary.KubectlNamespace)
+	kubectlHelpers.TeardownC8Helm(t, &secondary.KubectlNamespace)
 }
 
 func cleanupKubernetes(t *testing.T) {
@@ -369,7 +210,7 @@ func cleanupKubernetes(t *testing.T) {
 	k8s.RunKubectl(t, &secondary.KubectlSystem, "delete", "service", "internal-dns-lb")
 }
 
-// New stuff
+// Multi-Region Operational Procedure Additions
 
 func createElasticAWSSecret(t *testing.T) {
 	t.Log("[ELASTICSEARCH] Creating AWS Secret for Elasticsearch 🚀")
@@ -386,10 +227,10 @@ func createElasticAWSSecret(t *testing.T) {
 	S3AWSAccessKey := helpers.FetchSensitiveTerraformOutput(t, terraformOptions, "s3_aws_access_key")
 	S3AWSSecretAccessKey := helpers.FetchSensitiveTerraformOutput(t, terraformOptions, "s3_aws_secret_access_key")
 
-	helpers.RunSensitiveKubectlCommand(t, &primary.KubectlNamespace, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
-	helpers.RunSensitiveKubectlCommand(t, &secondary.KubectlNamespace, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
-	helpers.RunSensitiveKubectlCommand(t, &primary.KubectlFailover, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
-	helpers.RunSensitiveKubectlCommand(t, &secondary.KubectlFailover, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
+	kubectlHelpers.RunSensitiveKubectlCommand(t, &primary.KubectlNamespace, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
+	kubectlHelpers.RunSensitiveKubectlCommand(t, &secondary.KubectlNamespace, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
+	kubectlHelpers.RunSensitiveKubectlCommand(t, &primary.KubectlFailover, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
+	kubectlHelpers.RunSensitiveKubectlCommand(t, &secondary.KubectlFailover, "create", "secret", "generic", "elasticsearch-env-secret", fmt.Sprintf("--from-literal=S3_SECRET_KEY=%s", S3AWSSecretAccessKey), fmt.Sprintf("--from-literal=S3_ACCESS_KEY=%s", S3AWSAccessKey))
 
 	k8s.WaitUntilSecretAvailable(t, &primary.KubectlNamespace, "elasticsearch-env-secret", 5, 15*time.Second)
 	k8s.WaitUntilSecretAvailable(t, &secondary.KubectlNamespace, "elasticsearch-env-secret", 5, 15*time.Second)
@@ -414,49 +255,49 @@ func createElasticAWSSecret(t *testing.T) {
 func createElasticBackupRepoPrimary(t *testing.T) {
 	t.Log("[ELASTICSEARCH] Creating Elasticsearch Backup Repository 🚀")
 
-	helpers.ConfigureElasticBackup(t, primary, clusterName)
+	kubectlHelpers.ConfigureElasticBackup(t, primary, clusterName)
 }
 
 func createElasticBackupPrimary(t *testing.T) {
 	t.Log("[ELASTICSEARCH BACKUP] Creating Elasticsearch Backup 🚀")
 
-	helpers.CreateElasticBackup(t, primary, backupName)
+	kubectlHelpers.CreateElasticBackup(t, primary, backupName)
 }
 
 func checkThatElasticBackupIsPresentPrimary(t *testing.T) {
 	t.Log("[ELASTICSEARCH BACKUP] Checking if Elasticsearch Backup is present 🚀")
 
-	helpers.CheckThatElasticBackupIsPresent(t, primary, backupName)
+	kubectlHelpers.CheckThatElasticBackupIsPresent(t, primary, backupName)
 }
 
 func createElasticBackupRepoSecondary(t *testing.T) {
 	t.Log("[ELASTICSEARCH] Creating Elasticsearch Backup Repository 🚀")
 
-	helpers.ConfigureElasticBackup(t, secondary, clusterName)
+	kubectlHelpers.ConfigureElasticBackup(t, secondary, clusterName)
 }
 
 func checkThatElasticBackupIsPresentSecondary(t *testing.T) {
 	t.Log("[ELASTICSEARCH BACKUP] Checking if Elasticsearch Backup is present 🚀")
 
-	helpers.CheckThatElasticBackupIsPresent(t, secondary, backupName)
+	kubectlHelpers.CheckThatElasticBackupIsPresent(t, secondary, backupName)
 }
 
 func restoreElasticBackupSecondary(t *testing.T) {
 	t.Log("[ELASTICSEARCH BACKUP] Restoring Elasticsearch Backup 🚀")
 
-	helpers.RestoreElasticBackup(t, secondary, backupName)
+	kubectlHelpers.RestoreElasticBackup(t, secondary, backupName)
 }
 
 func deleteSecondaryRegion(t *testing.T) {
 	t.Log("[REGION REMOVAL] Deleting secondary region 🚀")
 
-	helpers.TeardownC8Helm(t, &secondary.KubectlNamespace)
+	kubectlHelpers.TeardownC8Helm(t, &secondary.KubectlNamespace)
 }
 
 func createFailoverDeploymentPrimary(t *testing.T) {
 	t.Log("[FAILOVER] Creating failover deployment 🚀")
 
-	helpers.InstallUpgradeC8Helm(t, &primary.KubectlFailover, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, false, true, false, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &primary.KubectlFailover, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, false, true, false, map[string]string{})
 
 	k8s.WaitUntilDeploymentAvailable(t, &primary.KubectlNamespace, "camunda-zeebe-gateway", 20, 15*time.Second)
 
@@ -471,7 +312,7 @@ func createFailoverDeploymentPrimary(t *testing.T) {
 func pointPrimaryZeebeToFailver(t *testing.T) {
 	t.Log("[FAILOVER] Pointing primary Zeebe to failover Elastic 🚀")
 
-	helpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, true, true, false, true, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, true, true, false, true, map[string]string{})
 
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlNamespace, "camunda-zeebe-3", 20, 15*time.Second)
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlNamespace, "camunda-zeebe-2", 20, 15*time.Second)
@@ -489,7 +330,7 @@ func recreateCamundaInSecondary(t *testing.T) {
 		"tasklist.enabled":                    "false",
 	}
 
-	helpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, false, false, false, true, setValues)
+	kubectlHelpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, false, false, false, true, setValues)
 
 	// expected that only 1 and 3 comes on
 	// 0 and 4 should be in not ready state
@@ -576,7 +417,7 @@ func installWebAppsSecondary(t *testing.T) {
 		"global.multiregion.installationType": "failBack",
 	}
 
-	helpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, setValues)
+	kubectlHelpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, setValues)
 
 	k8s.WaitUntilDeploymentAvailable(t, &secondary.KubectlNamespace, "camunda-operate", 20, 15*time.Second)
 	k8s.WaitUntilDeploymentAvailable(t, &secondary.KubectlNamespace, "camunda-tasklist", 20, 15*time.Second)
@@ -589,7 +430,7 @@ func pointC8BackToElastic(t *testing.T) {
 		"tasklist.enabled":                    "false",
 	}
 
-	helpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, true, true, false, false, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &primary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 0, true, true, false, false, map[string]string{})
 
 	k8s.RunKubectl(t, &primary.KubectlNamespace, "scale", "deployment", "camunda-operate", "--replicas=0")
 	k8s.RunKubectl(t, &primary.KubectlNamespace, "scale", "deployment", "camunda-tasklist", "--replicas=0")
@@ -599,16 +440,16 @@ func pointC8BackToElastic(t *testing.T) {
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlNamespace, "camunda-zeebe-1", 20, 15*time.Second)
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlNamespace, "camunda-zeebe-0", 20, 15*time.Second)
 
-	require.True(t, helpers.StatefulSetContains(t, &primary.KubectlNamespace, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
+	require.True(t, kubectlHelpers.StatefulSetContains(t, &primary.KubectlNamespace, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
 
-	helpers.InstallUpgradeC8Helm(t, &primary.KubectlFailover, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, true, true, true, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &primary.KubectlFailover, remoteChartVersion, remoteChartName, remoteChartSource, 0, false, true, true, true, map[string]string{})
 
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlFailover, "camunda-zeebe-1", 20, 15*time.Second)
 	k8s.WaitUntilPodAvailable(t, &primary.KubectlFailover, "camunda-zeebe-0", 20, 15*time.Second)
 
-	require.True(t, helpers.StatefulSetContains(t, &primary.KubectlFailover, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
+	require.True(t, kubectlHelpers.StatefulSetContains(t, &primary.KubectlFailover, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
 
-	helpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, setValuesSecondary)
+	kubectlHelpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, setValuesSecondary)
 
 	// 2 pods are sleeping indefinitely and block rollout
 	k8s.RunKubectl(t, &secondary.KubectlNamespace, "delete", "pod", "camunda-zeebe-0", "camunda-zeebe-1", "camunda-zeebe-2", "camunda-zeebe-3")
@@ -616,20 +457,20 @@ func pointC8BackToElastic(t *testing.T) {
 	k8s.WaitUntilPodAvailable(t, &secondary.KubectlNamespace, "camunda-zeebe-3", 20, 15*time.Second)
 	k8s.WaitUntilPodAvailable(t, &secondary.KubectlNamespace, "camunda-zeebe-1", 20, 15*time.Second)
 
-	require.True(t, helpers.StatefulSetContains(t, &secondary.KubectlNamespace, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
+	require.True(t, kubectlHelpers.StatefulSetContains(t, &secondary.KubectlNamespace, "camunda-zeebe", "http://camunda-elasticsearch-master-hl.camunda-secondary.svc.cluster.local:9200"))
 
 }
 
 func removeFailOverRegion(t *testing.T) {
 	t.Log("[FAILOVER] Removing failover region 🚀")
 
-	helpers.TeardownC8Helm(t, &primary.KubectlFailover)
+	kubectlHelpers.TeardownC8Helm(t, &primary.KubectlFailover)
 }
 
 func removeFailBackSecondary(t *testing.T) {
 	t.Log("[FAILOVER] Removing failback flag 🚀")
 
-	helpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, map[string]string{})
+	kubectlHelpers.InstallUpgradeC8Helm(t, &secondary.KubectlNamespace, remoteChartVersion, remoteChartName, remoteChartSource, 1, true, true, false, false, map[string]string{})
 
 	// 2 pods are sleeping indefinitely and block rollout
 	k8s.RunKubectl(t, &secondary.KubectlNamespace, "delete", "pod", "camunda-zeebe-0", "camunda-zeebe-1", "camunda-zeebe-2", "camunda-zeebe-3")
@@ -644,28 +485,28 @@ func checkTheMath(t *testing.T) {
 	t.Log("[MATH] Checking the math 🚀")
 
 	t.Log("[MATH] Checking if the primary deployment has even broker IDs")
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-0")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-1")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-2")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-3")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-0")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-1")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-2")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-3")))
 
 	t.Log("[MATH] Checking if the secondary deployment has odd broker IDs")
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-0")))
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-1")))
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-2")))
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-3")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-0")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-1")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-2")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &secondary.KubectlNamespace, "camunda-zeebe-3")))
 }
 
 func checkTheMathFailover(t *testing.T) {
 	t.Log("[MATH] Checking the math for Failover 🚀")
 
 	t.Log("[MATH] Checking if the primary deployment has even broker IDs")
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-0")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-1")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-2")))
-	require.True(t, helpers.IsEven(helpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-3")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-0")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-1")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-2")))
+	require.True(t, helpers.IsEven(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlNamespace, "camunda-zeebe-3")))
 
 	t.Log("[MATH] Checking if the failover deployment has odd broker IDs")
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &primary.KubectlFailover, "camunda-zeebe-0")))
-	require.True(t, helpers.IsOdd(helpers.GetZeebeBrokerId(t, &primary.KubectlFailover, "camunda-zeebe-1")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlFailover, "camunda-zeebe-0")))
+	require.True(t, helpers.IsOdd(kubectlHelpers.GetZeebeBrokerId(t, &primary.KubectlFailover, "camunda-zeebe-1")))
 }
